Soda 클래스 sugarContent Double로 수정

+++ VendingMachine/VendingMachine/Class/Soda/LemonlimeDrink.swift
@@ -11,7 +11,7 @@ import Foundation
 class LemonlimeDrink: Soda {
     private let lemonScentContent: Bool
     
-    init(brand: String, capacity: Int, price: Int, name: String, date: Date, sugarContent: Int, lemonScentContent: Bool) {
+    init(brand: String, capacity: Int, price: Int, name: String, date: Date, sugarContent: Double, lemonScentContent: Bool) {
         self.lemonScentContent = lemonScentContent
         
         super.init(brand: brand, capacity: capacity, price: price, name: name, date: date, sugarContent: sugarContent)
