+++ VendingMachine/VendingMachine/Class/Soda/Soda.swift
@@ -9,9 +9,9 @@
 import Foundation
 
 class Soda: Beverage {
-    private let sugarContent: Int
+    private let sugarContent: Double
     
-    init(brand: String, capacity: Int, price: Int, name: String, date: Date, sugarContent: Int) {
+    init(brand: String, capacity: Int, price: Int, name: String, date: Date, sugarContent: Double) {
         self.sugarContent = sugarContent
         
         super.init(brand: brand, capacity: capacity, price: price, name: name, date: date)
